pipeline: choose the stencil from the srate the run is ACTUALLY on
Adversarial audit finding, and a real bug in the selection logic.

--srate-internal overrides deltaT inside ILE (integrate_likelihood_extrinsic_
batchmode: `deltaT = deltaT_internal`), so it -- not --srate -- is the grid the
sub-sample stencil steps along.  It is appended to the ILE command line by
util_RIFT_pseudo_pipe.py and never passes through the helper, which contains
zero references to it.  Result: --srate-internal 32768 with srate 4096, fmax
1700 made the helper see fNyq/fmax = 1.20 and pick 'sinc', while the run was
really at 9.64, where this branch's own measurements put cubic 10-30x ahead.
--srate-internal >= 4x the data rate is a documented requirement for low-mass
runs, so this is live configuration space, not a corner.

Second route to the same defect: the helper emits --srate only under
--propose-ile-convergence-options, and ILE's own --srate default is 16384, not
4096 -- so on the branch where the CPU threshold is reachable the decision input
was off by 4x by construction.

Fixed with effective_srate_for_stencil(); pseudo_pipe forwards the internal rate
as an explicit decision input (NOT a second --srate-internal emission).  The test
asserts the naive and corrected paths choose DIFFERENT stencils for that case, so
the guard cannot decay into a no-op, and reads ILE's --srate default back out of
the driver source so the duplicated constant cannot rot silently.

Also from the audit:

* The helper accepted an unvalidated stencil name and concatenated it onto every
  generated ILE command line; a typo built and submitted a whole workflow, then
  killed each job separately at run time.  Validated at build time now.
* The legacy scalar path was handed opts.interpolate_time raw.  Once stencil
  NAMES became legal, '--interpolate-time nearest' would have switched that
  path's interpolation ON while meaning the opposite in NoLoop.  It gets a
  derived boolean.
* Both dispatchers ended in an unguarded `return cubic`, reinstating the silent
  wrong-stencil behaviour this work removes for anyone calling them directly.
  They raise.
* Two comments in _with_rotation and _freqresponse still claimed 'sinc' is
  rejected on GPU.  Untrue since Q_inner_sinc landed.
* cupy's astype copies even when the dtype already matches, so the sinc wrapper
  allocated a redundant (n_ex, 2a) float64 buffer -- ~20 MB transient per
  detector per call at n_chunk=1.6e5, on the resource that caps n_chunk.
* test_calmarg_stencil_gating runs its CPU arms without a GPU, so it joins the CI
  job; the commit that added it claimed CI coverage it did not have.

And one the audit did not raise, found while fixing the above: the flag now takes
a VALUE, so '--internal-ile-interpolate-time False' passes the STRING 'False',
which is truthy in Python -- every "off" spelling would have sailed past the
`if opts...:` guard and then been rejected as an unknown stencil.  Added
is_off_request(); the test asserts off / auto / stencil-name are disjoint.

Audit also confirmed, by direct measurement, that this branch does NOT perturb
existing results: the full NoLoop likelihood is bitwise identical to d904e72d for
'nearest' and 'cubic', CPU and GPU, n_cal=1 and n_cal=4, including
cal_method='fused'.

+++ MonteCarloMarginalizeCode/Code/RIFT/likelihood/Q_inner_product.py
@@ -165,7 +165,11 @@ def Q_inner_product_sinc_cupy(Q, A, start_indices, fractional_offsets, window_si
     tap_first = -halfwidth + 1
     assert _offsets.shape == (n_taps,), \
         "weight-matrix stencil width %r disagrees with 2*halfwidth=%d" % (_offsets.shape, n_taps)
-    tap_weights_d = cupy.ascontiguousarray(tap_weights.astype(cupy.float64))
+    # ascontiguousarray alone: _sinc_lanczos_weight_matrix already builds float64, and cupy's
+    # astype copies even when the dtype already matches (copy=True is its default).  At
+    # n_chunk=1.6e5 that extra (n_ex, 2a) float64 buffer is ~20 MB of transient device memory per
+    # detector per call, on the resource that already caps how large n_chunk can be.
+    tap_weights_d = cupy.ascontiguousarray(tap_weights)
 
     out = cupy.empty(
         (num_extrinsic_samples, window_size),
